fix: solve async self-invocation issue
